% Generated by build script
\name{drought_state}
\alias{drought_state}
\title{Get state-level drought severity statistics (percent area)}
\usage{
drought_state()
}
\arguments{
\item{state}{Two-letter state abbreviation (e.g., "CA", "TX") or FIPS code}
\item{start}{Start date as "M/D/YYYY" (e.g., "1/1/2024")}
\item{end}{End date as "M/D/YYYY" (e.g., "12/31/2024")}
}
\value{tibble: date, state, none, d0, d1, d2, d3, d4}
\description{Get state-level drought severity statistics (percent area)}
